% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calfprop_hist.R
\name{calfprop_hist}
\alias{calfprop_hist}
\title{Function for generating heifer proportion histogram}
\usage{
calfprop_hist(data_pop)
}
\arguments{
\item{data_pop}{input data}
}
\description{
Function for generating heifer proportion histogram
}
