% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CattleDenmark.R
\docType{data}
\name{CattleDenmark}
\alias{CattleDenmark}
\title{Data summarising cattle herds in Denmark}
\format{
A data frame with 6 rows and 14 variables:
\describe{
\item{HerdTypeID}{ID of herd type}
\item{etc}{etc}
}
}
\usage{
data(CattleDenmark)
}
\description{
A data frame containing summary statistics
}
\keyword{datasets}
