% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cattle_population.R
\name{generate_cattle_population}
\alias{generate_cattle_population}
\title{Generate cattle herd population function}
\usage{
generate_cattle_population(N, matrix_distr)
}
\arguments{
\item{matrix_distr}{}
}
\description{
Generate cattle herd population function
}
