% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heifprop_hist.R
\name{heifprop_hist}
\alias{heifprop_hist}
\title{Function for generating heifer proportion histogram}
\usage{
heifprop_hist(data_pop)
}
\arguments{
\item{data_pop}{input data}
}
\description{
Function for generating heifer proportion histogram
}
