% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herdsize_hist.R
\name{herdsize_hist}
\alias{herdsize_hist}
\title{Function for generating herd size histogram}
\usage{
herdsize_hist(data_pop)
}
\arguments{
\item{data_pop}{}
}
\description{
Function for generating herd size histogram
}
