% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herdtype_bar.R
\name{herdtype_bar}
\alias{herdtype_bar}
\title{Function for generating herd type bar chart}
\usage{
herdtype_bar(data_pop)
}
\arguments{
\item{data_pop}{input data}
}
\description{
Function for generating herd type bar chart
}
