% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_cross_table.R
\name{level_cross_table}
\alias{level_cross_table}
\title{Function for generating cross table for overall population}
\usage{
level_cross_table(data_level)
}
\arguments{
\item{data_level}{}
}
\description{
Function for generating cross table for overall population
}
