% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_type.R
\name{level_type}
\alias{level_type}
\title{Function for generating herd type level table}
\usage{
level_type(data_level)
}
\arguments{
\item{data_level}{}
}
\description{
Function for generating herd type level table
}
