% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leveltype_bar.R
\name{leveltype_bar}
\alias{leveltype_bar}
\title{Function for generating herd type level bar chart}
\usage{
leveltype_bar(data_level)
}
\arguments{
\item{data_level}{input data}
}
\description{
Function for generating herd type level bar chart
}
