% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampfreq_table.R
\name{sampfreq_table}
\alias{sampfreq_table}
\title{Function for generating sample frequency table}
\usage{
sampfreq_table(data_level)
}
\arguments{
\item{data_level}{}
}
\description{
Function for generating sample frequency table
}
