% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdubcontrol-package.R, R/sdubcontrol.R
\docType{package}
\name{sdubcontrol-package}
\alias{sdubcontrol-package}
\alias{_PACKAGE}
\alias{sdubcontrol}
\title{A Shiny Application Illustrating Salmonella Dublin Control}
\usage{
sdubcontrol(
  data_pop,
  prevHerd,
  prevAni,
  probSamp,
  seAni,
  spAni,
  nHeif,
  nCalf,
  nAdult,
  nVol,
  seClin,
  spClin,
  probClinInf,
  probClin,
  seMilk,
  spMilk,
  missSamp = FALSE,
  burnin = 3
)
}
\arguments{
\item{data_pop}{input population}

\item{prevHerd}{surveillance parameter}

\item{prevAni}{surveillance parameter}

\item{probSamp}{surveillance parameter}

\item{seAni}{surveillance parameter}

\item{spAni}{surveillance parameter}

\item{nHeif}{surveillance parameter}

\item{nCalf}{surveillance parameter}

\item{nAdult}{surveillance parameter}

\item{nVol}{surveillance parameter}

\item{seClin}{surveillance parameter}

\item{spClin}{surveillance parameter}

\item{probClinInf}{surveillance parameter}

\item{probClin}{surveillance parameter}

\item{seMilk}{surveillance parameter}

\item{spMilk}{surveillance parameter}

\item{burnin}{number of burnin cycles}
}
\description{
What the package does (one paragraph).
}
\details{
Some description of the package
}
\references{
Some article that you might want users to look at
}
\seealso{
\code{\link[neatpkg]{pkg_new}} for the function that created this package
}
\author{
\strong{Maintainer}: Alice Puk Skarbye \email{aps@sund.ku.dk}

Authors:
\itemize{
  \item Matt Denwood \email{md@sund.ku.dk}
}

}
\keyword{internal}
