% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdubcontrol.R
\name{sdubcontrol}
\alias{sdubcontrol}
\title{Function for generating surveillance output}
\usage{
sdubcontrol(
  data_pop,
  prevHerd,
  prevAni,
  probSamp,
  seAni,
  spAni,
  nHeif,
  nCalf,
  nAdult,
  nVol,
  seClin,
  spClin,
  probClinInf,
  probClin,
  seMilk,
  spMilk,
  missSamp = FALSE,
  burnin = 3
)
}
\arguments{
\item{burnin}{}
}
\description{
Function for generating surveillance output
}
