% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CattleDenmark.R
\docType{data}
\name{CattleDenmark}
\alias{CattleDenmark}
\title{Data summarising cattle herds in Denmark}
\format{
A data frame with 6 rows and 14 variables:
\describe{
\item{CalfProp_alpha}{}
\item{CalfProp_beta}{}
\item{CalfProp_mean}{}
\item{CalfProp_sd}{}
\item{CattleYears_meanlog}{}
\item{CattleYears_sdlog}{}
\item{Freq}{}
\item{HeifProp_alpha}{}
\item{HeifProp_beta}{}
\item{HeifProp_mean}{}
\item{HeifProp_sd}{}
\item{HerdTypeID}{}
\item{HerdTypeText}{}
\item{Prop}{}
}
}
\usage{
data(CattleDenmark)
}
\description{
A data frame containing summary statistics
}
\keyword{datasets}
